Routing completed
You can now log in ---> displays view based on role
As a student you can click on available tests ---> Goes to that page

Using sessionStorage we now save user_id based on log in, test_id based on test. This will make it easy for us to save stuff to database. We also store the index of a student in the studentlist(of the view), just to display the right student (index not the same as id)